mudado a quantidade para float no JFrame
- foi alterado na criação do produto para quando o usuário for informar
  a quantidade da materia prima, permitir valores decimais

+++ MoonlightApp/src/main/java/com/moonlight/moonlightapp/views/produtos/CriarProdutoView.java
@@ -179,7 +179,7 @@ public class CriarProdutoView extends javax.swing.JFrame {
             }
         ) {
             Class[] types = new Class [] {
-                java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.Double.class, java.lang.Integer.class
+                java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.Double.class, java.lang.Float.class
             };
             boolean[] canEdit = new boolean [] {
                 false, false, false, false, true
